refactor: moved QuerySearchParamNames into save view module
