+++ frontend/src/components/ExplorerCard/ExplorerCard.tsx
@@ -19,7 +19,6 @@ import {
 import axios from 'axios';
 import TextToolTip from 'components/TextToolTip';
 import { SOMETHING_WENT_WRONG } from 'constants/api';
-import { querySearchParams } from 'constants/queryBuilderQueryNames';
 import { useGetSearchQueryParam } from 'hooks/queryBuilder/useGetSearchQueryParam';
 import { useQueryBuilder } from 'hooks/queryBuilder/useQueryBuilder';
 import { useDeleteView } from 'hooks/saveViews/useDeleteView';
@@ -31,7 +30,11 @@ import { mapCompositeQueryFromQuery } from 'lib/newQueryBuilder/queryBuilderMapp
 import { useEffect, useState } from 'react';
 import { useCopyToClipboard } from 'react-use';
 
-import { ExploreHeaderToolTip, SaveButtonText } from './constants';
+import {
+	ExploreHeaderToolTip,
+	querySearchParams,
+	SaveButtonText,
+} from './constants';
 import MenuItemGenerator from './MenuItemGenerator';
 import SaveViewWithName from './SaveViewWithName';
 import {
