+++ frontend/src/hooks/queryBuilder/useGetSearchQueryParam.ts
@@ -1,4 +1,4 @@
-import { QuerySearchParamNames } from 'constants/queryBuilderQueryNames';
+import { QuerySearchParamNames } from 'components/ExplorerCard/constants';
 import useUrlQuery from 'hooks/useUrlQuery';
 import { useMemo } from 'react';
